simplellmLECL/requirements.txt
ipykernel
langchain
python-dotenv

langchain_community
langchain-ollama
langchain-classic
langchain-core


## for web scaping 
beautifulsoup4

## vector db
faiss-cpu


langchain_groq
langchain_openai


fastapi

uvicorn

langserve